backend_api/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-JWT-Extended==3.20.0
flask-marshmallow==0.10.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
marshmallow==3.0.0rc8
PyJWT==1.7.1
pyodbc==4.0.26